refactor(web): simplify resampler and move to audio_utils module
Move StatefulResampler to a dedicated audio_utils module and simplify it by
removing fallback logic. Since soxr is now a required dependency in
pyproject.toml, there's no need for ImportError handling or linear
interpolation fallback.

Changes:
- Create src/family_assistant/web/audio_utils.py with simplified StatefulResampler
- Remove fallback linear interpolation code (no longer needed)
- Remove try/except around soxr import (guaranteed to be available)
- Update asterisk_live_api.py to import from audio_utils
- Remove numpy import from asterisk_live_api.py (no longer used)

This keeps the asterisk endpoint focused on protocol handling while audio
processing utilities are properly separated.
